Crate score_four

Score Four Library for Rust

This is a simple Score Four board library for Rust.

Example

use score_four::{Board, BoardStatus};

let mut board = Board::new();
board.put(1); // put a bead in (0, 0) position
assert_eq!(board.status(), BoardStatus::Ongoing);

Structs

BitBoard
A bit-board for Score Four

Board
A representation of a score four board

Enums

BoardStatus
Status of a board, playing or lines are made

Color
Colors of players to play

Constants

EMPTY
An empty bitboard.

NUM_COLORS
number of colors